Add checks for reinitialising library.
This will cause the second initialisation attempt to load the
cache file. In doing so, we discover that cache loading on
non-32bit platforms didn't work -- fix that, too.

+++ src/rufl_init.c
@@ -1590,7 +1590,7 @@ rufl_code rufl_load_cache(void)
 	struct rufl_font_list_entry *entry;
 	struct rufl_character_set *charset = NULL, *cur_charset;
 	struct rufl_unicode_map *umap = NULL;
-	unsigned int num_umaps = 0;
+	size_t num_umaps = 0;
 
 	fp = rufl_open_cache("rb");
 	if (!fp)
